First pass at updating typing, internal as well as external.
